Adds email template for fallback
Much like the 'brand.html' template, this offers an optional template that the user can override so that when the fallback email is sent, it looks more in line with all the other emails that come from the Django application.

I realize that I could just implement this using the fallback methods already available to me, but this seemed like a simple patch that would be generally useful to other sites.